+++ test/phoenix_live_view/plug_test.exs
@@ -12,7 +12,6 @@ defmodule Phoenix.LiveView.PlugTest do
       |> Plug.Test.init_test_session(config[:plug_session] || %{})
       |> Plug.Conn.put_private(:phoenix_endpoint, Endpoint)
       |> Plug.Conn.put_private(:phoenix_live_view, [])
-      |> Map.put(:path_params, %{"id" => "123"})
 
     {:ok, conn: conn}
   end
@@ -20,7 +19,7 @@ defmodule Phoenix.LiveView.PlugTest do
   test "with no session opts", %{conn: conn} do
     conn = LiveViewPlug.call(conn, DashboardLive)
 
-    assert conn.resp_body =~ ~s(session: %{path_params: %{"id" => "123"}})
+    assert conn.resp_body =~ ~s(session: %{})
   end
 
   test "with existing #{LiveViewPlug.link_header()} header", %{conn: conn} do
@@ -29,17 +28,17 @@ defmodule Phoenix.LiveView.PlugTest do
       |> put_req_header(LiveViewPlug.link_header(), "some.site.com")
       |> LiveViewPlug.call(DashboardLive)
 
-    assert conn.resp_body =~ ~s(session: %{path_params: %{"id" => "123"}})
+    assert conn.resp_body =~ ~s(session: %{})
   end
 
   @tag plug_session: %{user_id: "alex"}
   test "with session opts", %{conn: conn} do
     conn =
       conn
-      |> Plug.Conn.put_private(:phoenix_live_view, session: [:path_params, :user_id])
+      |> Plug.Conn.put_private(:phoenix_live_view, session: [:user_id])
       |> LiveViewPlug.call(DashboardLive)
 
-    assert conn.resp_body =~ ~s(session: %{path_params: %{"id" => "123"}, user_id: "alex"})
+    assert conn.resp_body =~ ~s(session: %{user_id: "alex"})
   end
 
   test "with a container", %{conn: conn} do