+++ test/support/router.ex
@@ -19,7 +19,7 @@ defmodule Phoenix.LiveViewTest.Router do
 
     # router test
     live "/router/thermo_defaults/:id", DashboardLive
-    live "/router/thermo_session/:id", DashboardLive, session: [:path_params, :user_id]
+    live "/router/thermo_session/:id", DashboardLive, session: [:user_id]
     live "/router/thermo_container/:id", DashboardLive, container: {:span, style: "flex-grow"}
     live "/router/thermo_layout/:id", DashboardLive, layout: {Phoenix.LiveViewTest.AlternativeLayout , :layout}